+++ .bashrc.d/misc-aliases.sh
@@ -1,3 +1,5 @@
+#!/usr/bin/env bash
+
 alias cls='clear'
 alias dir='ls -alF'
 alias md='mkdir'