+++ .bashrc.d/misc-vars.sh
@@ -1,3 +1,4 @@
-PS1='\[\e[32m\]\u@\H \[\e[33m\]\w\[\e[0m\]\n\t \$ '
+#!/usr/bin/env bash
+
 export EDITOR=vim