+++ .bashrc.d/prompt.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+PS1='\[\e[32m\]\u@\H \[\e[33m\]\w\[\e[0m\]\n\t \$ '
+