+++ .bashrc.d/python-virtualenv.sh
@@ -1,4 +1,0 @@
-export PROJECT_HOME=~/projects
-export WORKON_HOME=~/.virtualenvs
-source /usr/bin/virtualenvwrapper.sh
-